% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ctds.R
\name{build_ctds}
\alias{build_ctds}
\title{Build structures for manipulating CTDS models}
\usage{
build_ctds(g, Xloc = NULL, coords = NULL, Vdir = NULL)
}
\arguments{
\item{g}{Either an \code{igraph} or an adjacency matrix that defines the 
neighborhood structure of a spatial domain with \eqn{n} locations.}

\item{Xloc}{\eqn{n x p} Matrix of location-based drivers of movement.
Each row defines covariates for a single location.  If \code{NULL}, then an 
intercept-only matrix will be automatically generated instead.}

\item{coords}{\eqn{n x d} Matrix associating coordinates with each location.
Used for determining covariates for directional drivers of motion.  
If \code{NULL}, then the returned \code{ctds} object will not be able to 
fit models with directional drivers of motion.}

\item{Vdir}{Either one \eqn{n x d} matrix, or a list of such matrices, 
each of which defines one directional driver of movement.  For example, 
the \code{j}th row in the \code{i}th \code{Vdir} matrix specifies a vector.  
The vector defines the strength and magnitude of the \code{i}th directional 
driver of movement at location \code{j}.}
}
\value{
A \code{ctds_struct} object containing information needed to compute 
  on the CTDS model.  One of the most important components of
  \code{ctds_struct} is the data.frame \code{edge_df}.  The data.frame is 
  structured so that all of the out-edges from a node appear as a consecutive 
  block of rows.
}
\description{
Build structures for manipulating CTDS models
}
\examples{
library(igraph)

# number of nodes (n) along each dimension (d)
n = 100
d = 2

# build 2D lattice in igraph with directed edges
g = make_lattice(dimvector = rep(n,d), directed = TRUE, mutual = TRUE)

# associate euclidean coordinates with lattice points
coords = expand.grid(x = 1:n, y = 1:n)

# directional driver of motion toward center of lattice
Vdir = t(apply(coords, 1, function(x) { x - rep(n/2, d) }))

# initialize structures for working with ctds models
ctds = build_ctds(g = g, Xloc = NULL, coords = coords, Vdir = Vdir)
}
