% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds.bsrw.R
\name{ctds.bsrw}
\alias{ctds.bsrw}
\title{Backward-sample a CTDS path without observations or params.}
\usage{
ctds.bsrw(n, ctds_struct, af, ff)
}
\arguments{
\item{n}{number of paths to sample}

\item{ctds_struct}{Output of \code{build_ctds}; a representation of a CTDS 
domain formatted for computing}

\item{af}{Final state distribution vector (sparse)}

\item{ff}{Forward-filtering distributions, which induces a path length.}
}
\value{
A vector of edges, representing a path
}
\description{
Backward-sampling follows via a random walk along the CTDS graph structure
}
