% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds.ffrw.R
\name{ctds.ffrw}
\alias{ctds.ffrw}
\title{Forward-filter a CTDS state distribution without observations or params.}
\usage{
ctds.ffrw(ctds_struct, a0, steps)
}
\arguments{
\item{ctds_struct}{Output of \code{build_ctds}; a representation of a CTDS 
domain formatted for computing}

\item{a0}{Initial state distribution vector (sparse)}

\item{steps}{Number of transitions to forward filter for}
}
\value{
A list \code{steps + 1} entries.  The initial state distribution 
  \code{a0} is contained in the first entry, and each successive entry has 
  the state distribution after one additional diffusion according to a 
  random walk along the spatial domain's graph, specified in 
  \code{ctds_struct}.
}
\description{
Forward-filtering follows via a random walk along the CTDS graph structure
}
