% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds.fwdsim.R
\name{ctds.fwdsim}
\alias{ctds.fwdsim}
\title{Forward simulation of a CTDS object}
\usage{
ctds.fwdsim(
  ctds_struct,
  beta_loc,
  beta_dir,
  v0,
  t0,
  tf,
  max.steps = Inf,
  beta_ar = NULL,
  v0.last = NULL
)
}
\arguments{
\item{ctds_struct}{Output of \code{build_ctds}; a representation of a CTDS 
domain formatted for computing}

\item{beta_loc}{coefficients for location-based drivers of movement}

\item{beta_dir}{coefficients for direction-based drivers of movement}

\item{v0}{the node from which sampling should start}

\item{t0}{the time at which sampling begins}

\item{tf}{the time after which sampling should end}

\item{max.steps}{the maximum number of transitions to simulate}

\item{beta_ar}{strength of directional-persistence parameter}

\item{v0.last}{the last-exited node before \code{v0}, if using directional
persistence---e.g., if \code{beta_ar != 0}.}
}
\description{
Forward-samples from the generating distributions for the CTDS model until 
either the object is observable at time \code{tf}, or a maximum number of 
transitions is realized (\code{max.steps}).
}
