% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds.observe.R
\name{ctds.observe}
\alias{ctds.observe}
\title{Simulate observing a complete CTDS trajectory at specified timepoints}
\usage{
ctds.observe(states, times, t.obs)
}
\arguments{
\item{states}{Complete record of states visited}

\item{times}{Times at which each of \code{states} was visited}

\item{t.obs}{Times at which the trajectory should be observed}
}
\description{
Given the components of a completely observed CTDS trajectory, extract the 
location of a trajectory at the target times \code{t.obs}.
}
