% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds.shortest_impute.R
\name{ctds.shortest_impute}
\alias{ctds.shortest_impute}
\title{Shortest-path CTDS imputation}
\usage{
ctds.shortest_impute(states, times, ctds_struct)
}
\arguments{
\item{states}{Sequence of nodes at which a CTDS is observed}

\item{times}{Times at which observations were recorded}

\item{ctds_struct}{Output of \code{build_ctds}; a representation of a 
CTDS domain formatted for computing}
}
\description{
This method returns a CTDS imputation that uses the shortest path to connect 
observations.  Since the exact in-edge is unknown, this will be randomly 
selected.  The method can be used to initialize latent state for Gibbs 
sampling.
}
