% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expocall_expm.R
\name{expocall_gpadm}
\alias{expocall_gpadm}
\title{Extract directed CTDS edges that begin and end at locations in locs}
\usage{
expocall_gpadm(H, t, nrows, ncols)
}
\arguments{
\item{x}{vector of states visited}
}
\description{
Extract directed CTDS edges that begin and end at locations in locs
}
