% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expokit_gpadm.R
\name{expokit_gpadm}
\alias{expokit_gpadm}
\title{Extract directed CTDS edges that begin and end at locations in locs}
\usage{
expokit_gpadm(ideg, m, t, H, ldh, wsp, lwsp, ipiv, iexph, ns, iflag)
}
\arguments{
\item{x}{vector of states visited}
}
\description{
Extract directed CTDS edges that begin and end at locations in locs
}
