% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_edges.R
\name{filter_edges}
\alias{filter_edges}
\title{Extract directed CTDS edges that begin and end at locations in locs}
\usage{
filter_edges(locs, inedges_by_loc, loc_start, fromlocs_by_edge)
}
\arguments{
\item{locs}{vector of locations for which edges should be filtered}

\item{inedges_by_loc}{edges sorted/grouped by their endpoint location}

\item{loc_start}{index of first in-edge for a location in 
\code{inedges_by_loc}}

\item{fromlocs_by_edge}{location from which each edge begins}
}
\value{
vector of edges that begin and end at locations in \code{locs}
}
\description{
Extract directed CTDS edges that begin and end at locations in locs
}
