% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intContains.R
\name{intContains}
\alias{intContains}
\title{Nimble version of a basic contains function}
\usage{
intContains(x, vec)
}
\arguments{
\item{x}{value to look for in \code{vec}}

\item{vec}{integer vector to search over}
}
\description{
Function returns TRUE if the integer vector \code{vec} contains the element 
\code{x}
}
