% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intWhich.R
\name{intWhich}
\alias{intWhich}
\title{Find the index at which \code{x} first appears in \code{vec}}
\usage{
intWhich(x, vec)
}
\arguments{
\item{x}{value to look for in \code{vec}}

\item{vec}{integer vector to search over}
}
\value{
the first index \code{ind} such that \code{vec[ind] == x}, or 0 if 
  \code{x} is not in \code{vec}
}
\description{
Find the index at which \code{x} first appears in \code{vec}
}
