% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mhrw1DAdaptive.R
\name{Mhrw1DAdaptive}
\alias{Mhrw1DAdaptive}
\title{Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class}
\description{
Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class

Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class
}
\details{
A-RWMH sampling for univariate densities, following 
Andrieu and Thoms (2008).  A-RWMH samplers need to maintain information 
about the target distribution being explored between sampler calls.  When 
A-RWMH samplers are used as a Gibbs step, required state is a nuisance to 
track and increases the likelihood for implementation errors.  This class is 
designed to make such samplers accessible by automatically/internally 
managing the state and required adaptation.
}
\references{
Andrieu, Christophe, and Johannes Thoms. "A tutorial on adaptive 
  MCMC." Statistics and computing 18.4 (2008): 343-373.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{Mhrw1DAdaptive$print()}}
\item \href{#method-new}{\code{Mhrw1DAdaptive$new()}}
\item \href{#method-sample}{\code{Mhrw1DAdaptive$sample()}}
\item \href{#method-clone}{\code{Mhrw1DAdaptive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Output the current values of the sampler's state, and adaptation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mhrw1DAdaptive$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Adaptive Random Walk Metropolis-Hastings sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mhrw1DAdaptive$new(x, sd, lp, C = 1, alpha = 0.44, adaptive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Initial location for sampler}

\item{\code{sd}}{Initial proposal standard deviation}

\item{\code{lp}}{Function to evaluate log of target density, given \code{x}}

\item{\code{C}}{Base step-size scale.  Note: Setting C=1 can lead to adaptation 
that occurs to quickly, yielding degenerate proposal covariance 
matrices.}

\item{\code{alpha}}{Target univariate acceptance rate for adaptation}

\item{\code{adaptive}}{\code{TRUE} to adapt proposal covariance \code{Sigma}, 
\code{FALSE} otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Use RWMH to propose, then accept/reject, a new value from the target 
distribution.  If the sampler was configured with \code{adaptive=TRUE}, 
then the sampler will automatically adapt as well.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mhrw1DAdaptive$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to pass to the function that evaluates
the log-posterior}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mhrw1DAdaptive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
