% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MhrwAdaptive.R
\name{MhrwAdaptive}
\alias{MhrwAdaptive}
\title{Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class}
\description{
Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class

Adaptive Random Walk Metropolis-Hastings (A-RWMH) sampler class
}
\details{
A-RWMH sampling for multivariate densities, following Algorithm 6 in 
Andrieu and Thoms (2008).  A-RWMH samplers need to maintain information 
about the target distribution being explored between sampler calls.  When 
A-RWMH samplers are used as a Gibbs step, required state is a nuisance to 
track and increases the likelihood for implementation errors.  This class is 
designed to make such samplers accessible by automatically/internally 
managing the state and required adaptation.
}
\references{
Andrieu, Christophe, and Johannes Thoms. "A tutorial on adaptive 
  MCMC." Statistics and computing 18.4 (2008): 343-373.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{MhrwAdaptive$print()}}
\item \href{#method-new}{\code{MhrwAdaptive$new()}}
\item \href{#method-sample}{\code{MhrwAdaptive$sample()}}
\item \href{#method-clone}{\code{MhrwAdaptive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Output the current values of the sampler's state, and adaptation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MhrwAdaptive$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a new Adaptive Random Walk Metropolis-Hastings sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MhrwAdaptive$new(
  x,
  mu,
  Sigma,
  lambda,
  lp,
  C = 0.75,
  alpha = 1,
  alpha_star = 0.44,
  adaptive = TRUE,
  adaptation_frequency = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Initial location for sampler}

\item{\code{mu}}{Initial guess of target distribution's mean}

\item{\code{Sigma}}{Initial guess of target distribution's covariance matrix}

\item{\code{lambda}}{Initial scaling factors, for adaptation}

\item{\code{lp}}{Function to evaluate log of target density, given \code{x}}

\item{\code{C}}{Base step-size scale.  Note: Setting C=1 can lead to adaptation 
that occurs to quickly, yielding degenerate proposal covariance 
matrices.}

\item{\code{alpha}}{Power of the diminishing adaptation step size's decay.
alpha must be in the range (1/(1+lambda), 1]}

\item{\code{alpha_star}}{Target univariate acceptance rate for adaptation}

\item{\code{adaptive}}{\code{TRUE} to adapt proposal covariance \code{Sigma}, 
\code{FALSE} otherwise.}

\item{\code{adaptation_frequency}}{How often to adapt sampler parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\subsection{Method \code{sample()}}{
Use RWMH to propose, then accept/reject, a new value from the target 
distribution.  If the sampler was configured with \code{adaptive=TRUE}, 
then the sampler will automatically adapt as well.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MhrwAdaptive$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to pass to the function that evaluates
the log-posterior}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MhrwAdaptive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
