% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctds_observations.R
\name{plot.ctds_observations}
\alias{plot.ctds_observations}
\title{Plot observations of a CTDS object}
\usage{
plot.ctds_observations(
  x,
  ctds_struct,
  pt_size = 0.1,
  pt_col = "grey60",
  jitter.factor = 0,
  ctds_realization = NULL,
  realization_alpha = 0.55,
  ...
)
}
\arguments{
\item{x}{\code{ctds_observation} object for plotting}

\item{ctds_struct}{\code{ctds_struct} object with domain information to 
assist with plotting}

\item{pt_size}{When plotting a spatial domain defined via point locations, 
\code{pt_size} controls the size of the points in the grid}

\item{pt_col}{When plotting a spatial domain defined via point locations, 
\code{pt_size} controls the color of the points in the grid}

\item{jitter.factor}{CTDS realizations may be plotted with their coordinates 
jittered slightly to aid in readability.  This argument is a number that
specifies the \code{factor} argument to \code{base::jitter}.}

\item{ctds_realization}{\code{ctds_realization} object containing 
observations of the \code{ctds_realization} to plot, or a list of such 
objects to plot.}

\item{realization_alpha}{amount of transparency with which to underlay the 
\code{ctds_realization} object}

\item{...}{Additional plotting arguments}
}
\description{
Method allows for underlaying complete trajectories
}
