% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.gumbeltrick.R
\name{sample.gumbeltrick}
\alias{sample.gumbeltrick}
\title{Sample a categorical variable when given log-probabilities}
\usage{
sample.gumbeltrick(log.p)
}
\arguments{
\item{log.p}{log of categorical probabilitites; the probabilities do not 
need to be normalized}
}
\description{
Sample a categorical variable when given log-probabilities
}
\references{
https://timvieira.github.io/blog/post/2014/07/31/gumbel-max-trick/
}
