% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctds_nbhd_ll.R, R/ctds_obs_lik.R
\name{ctds_nbhd_ll}
\alias{ctds_nbhd_ll}
\alias{ctds_obs_lik}
\title{Extract directed CTDS edges that begin and end at locations in locs}
\usage{
ctds_nbhd_ll(
  x,
  durations,
  N,
  inedges_by_loc,
  inloc_start,
  tolocs_by_edge,
  fromlocs_by_edge,
  outedges_by_loc,
  loc_start,
  Xloc,
  betaLoc,
  Xdir,
  betaDir,
  W,
  betaAR,
  nbrlocs_by_loc,
  nbrlocs_start,
  log = 0
)

ctds_obs_lik(x, durations, N, inedges_by_loc, inloc_start, log = 0)
}
\arguments{
\item{x}{vector of states visited}
}
\description{
Extract directed CTDS edges that begin and end at locations in locs
}
